requirements.txt
pillow==10.3.0
lxml==5.2.2
XlsxWriter==3.2.0
python-pptx==0.6.23
gpt4all==2.7.0
Flask==2.2.5
Flask-Cors==4.0.0
streamlit==1.34.0
langchain_openai==0.3.35
langgraph==0.6.11
pypdf==6.8.0
python-docx==1.2.0
